+++ cibuildwheel/selector.py
@@ -1,6 +1,6 @@
 import itertools
 from dataclasses import dataclass
-from enum import Enum
+from enum import StrEnum
 from fnmatch import fnmatch
 from typing import Any
 
@@ -24,7 +24,7 @@ def selector_matches(patterns: str, string: str) -> bool:
     return any(fnmatch(string, pat) for pat in expanded_patterns)
 
 
-class EnableGroup(Enum):
+class EnableGroup(StrEnum):
     """
     Groups of build selectors that are not enabled by default.
     """
@@ -32,6 +32,10 @@ class EnableGroup(Enum):
     CPythonFreeThreading = "cpython-freethreading"
     CPythonPrerelease = "cpython-prerelease"
     PyPy = "pypy"
+
+    @classmethod
+    def all_groups(cls) -> frozenset["EnableGroup"]:
+        return frozenset(cls)
 
 
 @dataclass(frozen=True, kw_only=True)