+++ test/test_container_images.py
@@ -11,7 +11,7 @@ dockcross_only_project = test_projects.new_c_project(
         import os
 
         # check that we're running in the correct docker image as specified in the
-        # environment options CIBW_MANYLINUX1_*_IMAGE
+        # environment options CIBW_MANYLINUX_*_IMAGE
         if "linux" in sys.platform and not os.path.exists("/dockcross"):
             raise Exception(
                 "/dockcross directory not found. Is this test running in the correct docker image?"
@@ -43,9 +43,12 @@ def test(tmp_path):
     )
 
     # also check that we got the right wheels built
+    manylinux_versions = ["manylinux_2_5", "manylinux1", "manylinux_2_17", "manylinux2014"]
     expected_wheels = [
         w
-        for w in utils.expected_wheels("spam", "0.1.0", musllinux_versions=[])
+        for w in utils.expected_wheels(
+            "spam", "0.1.0", manylinux_versions=manylinux_versions, musllinux_versions=[]
+        )
         if "-cp38-" in w or "-cp39-" in w
     ]
     assert set(actual_wheels) == set(expected_wheels)