+++ test/test_environment.py
@@ -89,7 +89,7 @@ def test_overridden_path(tmp_path, capfd):
                 },
             )
 
-    assert len(os.listdir(output_dir)) == 0
+    assert len(list(output_dir.iterdir())) == 0
     captured = capfd.readouterr()
     assert "python available on PATH doesn't match our installed instance" in captured.err.replace(
         "venv", "installed"